requirements.txt
install==1.3.5
lxml==4.9.4
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
pytesseract==0.3.10
python-docx==1.1.0
typing_extensions==4.9.0
